atomic desain for comment feature
